feat: add Optional type
This commit adds a new Optional type in the 'nulls' package to hold nullable values. The Optional type offers functionality for JSON Marshalling and Unmarshalling. Currently, the 'Scan' and 'Value' methods are not supported and will return an error. Corresponding tests for the Optional type were also added for coverage.